+++ backend/src/poi.rs
@@ -1,6 +1,12 @@
+use actix_web::{
+    HttpResponse, Responder, get,
+    web::{Data, Path},
+};
 use geozero::{ToWkt, wkb};
 use serde::{Serialize, ser::SerializeStruct};
 use sqlx::FromRow;
+
+use crate::AppState;
 
 #[derive(Debug, FromRow)]
 pub struct Poi {
@@ -57,9 +63,36 @@ impl Serialize for Poi {
     }
 }
 
-#[derive(Serialize, Debug, FromRow)]
-pub struct Brand {
-    pub id: i32,
-    pub name: String,
-    pub wikidata_id: Option<String>,
+#[get("/poi/{id}")]
+async fn get_poi_by_id(state: Data<AppState>, path: Path<i32>) -> impl Responder {
+    let id = path.into_inner();
+    match sqlx::query_as::<_, Poi>("SELECT * FROM poi WHERE id = $1")
+        .bind(id)
+        .fetch_one(&state.db)
+        .await
+    {
+        Err(why) => {
+            HttpResponse::NotFound().body(format!("No poi found with id: {}, error: {}", id, why))
+        }
+        Ok(poi) => HttpResponse::Ok().json(poi),
+    }
+}
+
+#[get("/poi/random/{count}")]
+async fn get_random_pois(state: Data<AppState>, path: Path<i64>) -> impl Responder {
+    let limit = path.into_inner();
+    let max_limit = 15;
+    if limit > 15 {
+        return HttpResponse::BadRequest().body(format!("Limit must be less than {}", max_limit));
+    }
+    match sqlx::query_as::<_, Poi>("SELECT * FROM poi LIMIT $1")
+        .bind(limit)
+        .fetch_all(&state.db)
+        .await
+    {
+        Err(why) => {
+            HttpResponse::NotFound().body(format!("Error while getting random POIs: {}", why))
+        }
+        Ok(pois) => HttpResponse::Ok().json(pois),
+    }
 }
